Propagate compiler flags for release builds, use release builds on Windows.
We set the build type on Windows because we assume we're building with mingw.
No need to consider multi-config generators because R doesn't use them anyway.

We propagate the compiler flags into the CMAKE_C(XX)_FLAGS_RELEASE variable
to avoid clobbering R's flags with Cmake's defaults for the Release build.